Extract logic to separate function
